requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.18.0
highspy>=1.7.0
requests>=2.28.0
beautifulsoup4>=4.12.0
scikit-learn>=1.3.0
fpdf2>=2.7.0
supabase>=2.0.0
streamlit-analytics2>=0.10.0
pyyaml>=6.0.0
